+++ protected/display_accounts.php
@@ -25,6 +25,7 @@ include_once('/etc/geni-ar/settings.php');
 require_once('ldap_utils.php');
 require_once('db_utils.php');
 require_once('ar_constants.php');
+require_once('response_format.php');
 
 global $base_dn;
 global $acct_manager_url;
@@ -92,9 +93,9 @@ foreach ($accts as $acct) {
 }
 print '</tbody></table></div>';
 
-$sql = "SELECT * FROM " . $AR_TABLENAME . " WHERE request_state='DELETED'";
+$sql = "SELECT * FROM " . $AR_TABLENAME . " WHERE request_state='" . AR_STATE::DELETED . "'";
 $result = db_fetch_rows($sql);
-if ($result['code'] != 0) {
+if ($result[RESPONSE_ARGUMENT::CODE] != RESPONSE_ERROR::NONE) {
   process_error("Postgres database query failed");
   exit();
 }
@@ -119,7 +120,7 @@ foreach ($rows as $row) {
   $action_ts="";
   $sql = "SELECT performer, action_ts from idp_account_actions WHERE uid='" . $uname . "' and action_performed='Account Deleted' ORDER BY id desc";
   $action_result = db_fetch_rows($sql);
-  if ($result['code'] != 0) {
+  if ($result[RESPONSE_ARGUMENT::CODE] != RESPONSE_ERROR::NONE) {
     process_error("Postgres database query failed");
     exit();
   }