+++ CHANGES.md
@@ -2,6 +2,11 @@
 
 # [Release 1.8](https://github.com/GENI-NSF/geni-ar/milestones/1.8)
 
+* Removed `Current Reqs` tab; default is `Email Confirmed`.
+  Added an EMAIL_CONFIRMED state constant, and make more use
+  of constants. New requests change to state `CONFIRM`
+  once the automatic email to confirm their address is sent.
+  ([#148](https://github.com/GENI-NSF/geni-ar/issues/148))
 * Add initial system documentation
   ([#152](https://github.com/GENI-NSF/geni-ar/issues/152))